+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.3.2] - Upcoming
 
+### Added
+
+- Added `deleteAccessTokens` and `deleteRefreshTokens` methods in the `UserTokens` model
+
 ### Changed
 
 - Updated model meta to not dot arrays